clean: return type

+++ simple-array-sum/jl.jl
@@ -1,4 +1,4 @@
-simple_array_sum(ar::Array{Int}) = sum(ar)
+simple_array_sum(ar::Array{Int})::Int = sum(ar)
 
 n = parse(Int, readline(stdin))
 ns = map(str->parse(Int, str), split(readline(stdin), r"\s"))
